examples/DHT_sensor/requirements.txt
rpi-lgpio == 0.6
